Theme Manager
Basic first pass at the theme manager for the app. This will let us
define multiple themes and switch between them at will. This is a
departure from prior apps where the theme was based on the platform and
platform only. This can still be the case, but the theme itself isn’t
tied to the platform.